requirements.txt
aws-cdk-lib>=2.224.0  # Required for aws_s3vectors native CloudFormation support
constructs>=10.0.0,<11.0.0
aws_cdk.aws_lambda_python_alpha>=2.224.0a0
# Note: aws_cognito_identitypool_alpha was graduated to stable in aws-cdk-lib.aws_cognito_identitypool
boto3
PyYAML
cdk_nag
python-dotenv
pydantic
Jinja2>=3.1.5
aws-lambda-powertools
click
xmltodict>=0.13.0
